Completion of all implementation. Tested successfully.
Added in PtrDetails, from the examples during lesson3,
as the gc_details. This class contains details about a pointer.
The implementation of a list which stores multiple PtrDetails object,
or details on multiple pointers, is inside gc_pointer.h. An iterator,
to iterate through the list is inside gc_iterator.h.